requirements.txt
fastapi==0.68.1
python-multipart==0.0.5
uvicorn==0.15.0
numpy==1.23.5
python-dotenv==0.19.0
gunicorn==20.1.0
opencv-python-headless==4.6.0.66
deepface==0.0.79
tensorflow-cpu==2.13.0
scipy==1.10.1
psutil==5.9.5